Correction : le type "int" avait été utilisé par erreur pour le poids d'un terme à certains endroits à la place de "double".

+++ src/index/IndexHash.java
@@ -52,7 +52,7 @@ public class IndexHash extends Index {
 	}
 
 	@Override
-	public void setPoids(String terme, String urlDocument, int poids) {
+	public void setPoids(String terme, String urlDocument, double poids) {
 		index.get(terme).get(urlDocument).setPoids(poids);
 	}
 }
